anatomy/RAG/requirements.txt
ollama
faiss-cpu
numpy
transformers
torch